Fix all program for the presentation day
Now there is a file `experiments.txt` with the information
of how many lines on each file was spend on each part of the
project. These parts are divided in 3:

1. Ontology: Knowledge about the vocabulary
2. Model: The design of the reasoning model to answer the question
3. Parser: The code for the matcher that actually parses the text

Also the jSpy symlink was updated to a submodule in a stable
version.

The `.spy` files were adapted to work with this stable version.